refactor: change pkg folder and func name
